♻️ refactor : entitiy => entity
