+++ src/main/java/com/example/momentix/domain/reservation/service/ReservationService.java
@@ -8,8 +8,8 @@ import com.example.momentix.domain.events.repository.EventPlaceRepository;
 import com.example.momentix.domain.events.repository.EventsRepository;
 import com.example.momentix.domain.events.repository.eventtimes.EventTimesRepository;
 import com.example.momentix.domain.reservation.dto.ReservationResponseDto;
-import com.example.momentix.domain.reservation.entitiy.ReservationStatusType;
-import com.example.momentix.domain.reservation.entitiy.Reservations;
+import com.example.momentix.domain.reservation.entity.ReservationStatusType;
+import com.example.momentix.domain.reservation.entity.Reservations;
 import com.example.momentix.domain.reservation.repository.ReservationRepository;
 import com.example.momentix.domain.users.entity.Users;
 import com.example.momentix.domain.users.repository.UserRepository;
